requirements.txt
Flask==2.2.2
Pillow==9.3.0
Werkzeug==2.2.2

